First Commit -- Model Fits for Mechanical Data
Transferred over necessary data (for Liao et al. 2025 Sci Data) for the mechanical analysis presented in this paper. Modified code from original paper to plot model fits, as well as recalculate the corrected stress and strain data (accounting for to region) for the rigid resin data (not originally saved in the dataset)